feat: It reports!
I obviously spent a considerable amount of time before this working on
things to get them to this point. I felt like enough of the architecture
was in place and I could see some clear refactors to make things more of
a proper application, especially to cleanup state and some async abuses,
so here we go.