refact: Complete overhaul of data handling.
This maintains in memory copies of price DataFrames for the stocks by polling
the price directory for changes and then reloading the files in a separate
thread. This removes the price loading from stock-view-model building, which in
turn means the UI and various other places need to handle some incomplete or
inbetween states as things are loading.

This also marks the start of rendering OHLC of candle data using the CSV rather
than the JSON. We still write those, and can't get rid of the parsing code
because we use that to append to the data frames.